Buggy, but can now change cell in our table
Using new api endpoint, we can select a cell, and replace the value in it by selecting it and typing new text. Once a user is no longer focused on the cell (done typing), We use the endpoint to replace the corresponding cell in the database. Currently does not update value in database if our string has spaces.