new: endpoint to change field name, changed: endpoint that sends table field names
The webpage will need to be able to allow users to change the names of
their tables fields, therefore i created an endpoint that will allow for that. In addition, every table record has a unique key when created, but i do not think we need to send this key to the webpage through the /api/table-fields endpoint. The unique key was created to help in changing the value of a certain cell in our table using the recently created /api/change-cell endpoint. We are still sending this id field in our /api/get-table endpoint. This results in table in webpage not showing the id field.