Update index.js
LIne 44: double Quotes are used so we can have spaces in our table names. We also need to add the double quotes in select statement to preserve case sensitivity. or we get errors when looking for a table, saying that it doesnt exist.
Changed name of unique key "id" when creating a record to "unique_record_id" so that users can create a field named "id" without causing any errors.