Fixed Bug
Error: relation "movie" does not exist

when attempting to change the cell of a table that has a space in its name like "Hike club". Had to update the Query to include double quotes around the table name to ensure it works with spaces. Also added double quotes around fieldName for the same reason.