Switch case for input types. Checbox now works !!
It's a smart move to use a switch case because it lets me easily display various input types based on what I need. Take checkboxes, for instance—they need a different setup compared to other input types. So, with the switch case, I can tailor the configuration of the input element specifically for checkboxes, making it straightforward to handle the differences in rendering.