Added functionality to checkbox filters. Added comments
By using an object instead of an array, we can have the unique field that applied the filter be the key, and the filter function be the value. This allows us to remove the filter for the specific field when needed, because all we need is the key of the field whos filter is to be removed. When a filter is removed, the key value pair is removed from the object, and we reapply the remaining filters to the unmodified rows, which we get by making a request to the db (so that we get most recently updated state)