Fixed filtering bug
There was a bug where if you chose a filter, then tried to change it, the table would not be accurate This is because useEffect changes state of rows whenever a new filter is added. But if we tried to change the filter, it would fail because rows still has result of previous filter. so we must get fresh row data even when new filter is applied, not just when removing filters.