Very Important bug fixes. Read
Because we omitted the encodeURI step, the API received no trailing spaces, making it impossible to locate tables starting or ending with such spaces. I apply the trim function to eliminate spaces at the beginning and end of strings, along with encoding. While it might be redundant to use both, they collectively address the issue. This adjustment is essential for both getRows and getFields.

Additionally, my initial approach to filtering, using `record[field]` to ensure the presence of a value, proved less effective than anticipated. Notably, the unchecked filter for checkboxes, despite having columns with unchecked boxes, failed to display any rows. This discrepancy arises from the usage of true and false values for checkboxes. When utilizing `record[field]` to check for values, false is returned for unchecked checkboxes, resulting in nothing being rendered.